% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{antilogit}
\alias{antilogit}
\title{Anti-logit Transformation (Inverse Logit)}
\usage{
antilogit(x)
}
\arguments{
\item{x}{A numeric value to transform (can be any real number)}
}
\value{
A numeric value between 0 and 1 representing a probability
}
\description{
Transforms a real number to a probability using the anti-logit (logistic) function.
This is the inverse of the logit transformation, converting log-odds back to probabilities.
The function computes exp(x) / (1 + exp(x)).
}
