% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{arrow_in}
\alias{arrow_in}
\title{Build Arrow Expression for membership filters}
\usage{
arrow_in(field, values)
}
\arguments{
\item{field}{Character scalar column name or an arrow::Expression.}

\item{values}{Vector of values to match.}
}
\value{
An arrow::Expression suitable for use in read_parquet_dt().
}
\description{
Build Arrow Expression for membership filters
}
\examples{
\dontrun{
arrow_in("sex", c("men", "women"))
read_parquet_dt(
  "./data/myfile.parquet",
  filter = arrow_in("sex", "men")
)
}
}
