% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clone_dt}
\alias{clone_dt}
\title{Clone a data.table}
\usage{
clone_dt(dt, times, idcol = TRUE)
}
\description{
\code{clone_dt} replicates a data.table and binds the replicates a the bottom of
the original one. It also creates an column named\code{idcol} to identify each
iteration.
}
