% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{csv_to_fst}
\alias{csv_to_fst}
\title{convert csv to fst and optionally delete csv. It accepts vectors}
\usage{
csv_to_fst(csv_files, compression = 100L, delete_csv = FALSE)
}
\arguments{
\item{csv_files}{A character vector containing the file paths of the CSV files to be converted.}

\item{compression}{An integer specifying the compression level for the FST format. Default is 100.}

\item{delete_csv}{If TRUE, delete the original CSV files after conversion. Default is FALSE.}
}
\value{
Invisible NULL.
}
\description{
This function converts a list of CSV files to the FST format with optional compression.
}
\examples{
\dontrun{
files <- c("data/file1.csv", "data/file2.csv")
csv_to_fst(files, compression = 80L, delete_csv = TRUE)
}

}
