% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pss_gnpr_prbs.R
\name{dfr2}
\alias{dfr2}
\title{Round numeric matrix to two decimal places and convert to data frame for
display}
\usage{
dfr2(mat)
}
\arguments{
\item{mat}{Numeric matrix}
}
\value{
Data frame rounded to two decimal places
}
\description{
Round numeric matrix to two decimal places and convert to data frame for
display
}
