% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{do_cols_dt}
\alias{do_cols_dt}
\title{Selected columns in a datatable are added, subtracted etc, or a fn is applied to them}
\usage{
do_cols_dt(dt, cols_to_add, symbol = c(",", "+", "-", "*", "/"), fn = NULL)
}
\arguments{
\item{dt}{The data table on which operations are to be performed.}

\item{cols_to_add}{A character vector of column names to be combined in the operation.}

\item{symbol}{A character vector specifying the symbols to be used in combining columns. Default is c(",", "+", "-", "*", "/").}

\item{fn}{An optional function to apply to the combined result.}
}
\value{
A new data table containing the result of the column-wise operations.
}
\description{
This function performs column-wise operations on a data table, combining specified columns using given symbols.
}
\examples{
dt <- data.table(A = c(1, 2, 3), B = c(4, 5, 6), C = c(7, 8, 9))
do_cols_dt(dt, cols_to_add = c("A", "B", "C"), symbol = "+", fn = "sqrt")

}
