% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estim_beta_params}
\alias{estim_beta_params}
\title{`estim_beta_params` estimates the beta parameters from a mean and a variance given}
\usage{
estim_beta_params(mu, var)
}
\arguments{
\item{mu}{An integer between 0 and 1}

\item{var}{A positive integer}
}
\value{
A list made of the beta and alpha parameters calculated, called shape1 and shape2
}
\description{
`estim_beta_params` estimates the beta parameters from a mean and a variance given
}
\examples{
estim_beta_params(0.6, 5) #6.0006e-05   4.0004e-05
}
