% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fclamp_int}
\alias{fclamp_int}
\title{Clamp an Integer Vector Within a Range}
\usage{
fclamp_int(x, a = 0L, b = 1L, inplace = FALSE)
}
\arguments{
\item{x}{An integer vector to be clamped.}

\item{a}{The lower bound (default is 0).}

\item{b}{The upper bound (default is 1).}

\item{inplace}{Logical flag indicating whether to modify the input vector in place (default is false).}
}
\value{
An integer vector with values clamped to the range [a, b].
}
\description{
This function clamps the values of an integer vector to lie within a specified range [a, b].
Values below a are set to a and values above b are set to b. Optionally, the operation can be performed in-place.
}
