% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fct_to_int_cpp}
\alias{fct_to_int_cpp}
\title{Convert Factor to Integer (C++ Version)}
\usage{
fct_to_int_cpp(x, inplace = FALSE)
}
\arguments{
\item{x}{A factor object.}

\item{inplace}{Logical flag indicating whether to modify the input in place (default: false).}
}
\value{
An integer vector containing the underlying integer codes without factor attributes.
}
\description{
This function converts a factor (an integer vector with a "levels" attribute)
to its underlying integer representation by stripping away the factor's levels and class attributes.
By default, the function does not modify the input by reference (a copy is made), but if `inplace`
is set to `true`, the function will modify the input directly.
}
