% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fequal}
\alias{fequal}
\title{Check Numeric Vector Equality Within Tolerance}
\usage{
fequal(x, tol)
}
\arguments{
\item{x}{A numeric vector to test for equality.}

\item{tol}{A tolerance value. Differences below or equal to this value are considered equal.}
}
\value{
A logical value: TRUE if all non-missing elements are equal within the tolerance,
        and FALSE otherwise.
}
\description{
This function checks whether all non-missing elements in a numeric vector are equal
within a specified tolerance. Differences between elements that are less than or equal
to the tolerance are considered negligible.
}
