% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fget_C}
\alias{fget_C}
\title{Get Normalizing Constant for DPO Distribution}
\usage{
fget_C(x, mu, sigma)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles}

\item{mu}{vector of positive means}

\item{sigma}{vector of positive dispersion parameters}
}
\value{
Vector of log normalizing constants
}
\description{
Computes the logarithm of the normalizing constant for the DPO distribution.
This is an internal function used by the DPO distribution functions.
}
\details{
This function computes the logarithm of the normalizing constant required
for the DPO (Double Poisson) distribution. The computation follows the
algorithm from gamlss.dist but with optimizations for performance.
}
\note{
This function is based on the DPO distribution implementation from
the \pkg{gamlss.dist} package by Mikis Stasinopoulos, Robert Rigby,
and colleagues. The original gamlss.dist implementation is acknowledged
with gratitude.
}
\references{
Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for 
location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Rigby, R. A., Stasinopoulos, D. M., Heller, G. Z., and De Bastiani, F. (2019)
Distributions for modelling location, scale, and shape: Using GAMLSS in R, Chapman and Hall/CRC.
}
\seealso{
\code{\link{fdDPO}}, \code{\link{fpDPO}}, \code{\link{fqDPO}}
}
\author{
Chris Kypridemos (optimised implementation), based on original work by 
Bob Rigby and Mikis Stasinopoulos from gamlss.dist package
}
