% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pss_gnpr_prbs.R
\name{find_ale_frqs}
\alias{find_ale_frqs}
\title{Find allele frequencies}
\usage{
find_ale_frqs(ind_gts)
}
\arguments{
\item{ind_gts}{Genotypes in 2 x L x n_individuals arrays, representing two
binary SNPs at each locus for each individual, excluding repeated samples
of the same individual in different surveys}
}
\value{
2 x L matrix representing the frequencies of 0 and 1-coded SNP
alleles at each locus
}
\description{
Find allele frequencies
}
