% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_exp_ns_kps.R
\name{find_exp_ns_kps_smpd}
\alias{find_exp_ns_kps_smpd}
\title{Find expected numbers of kin-pairs between sampled animals}
\usage{
find_exp_ns_kps_smpd(exp_ns_kps_cmbd, k, p)
}
\arguments{
\item{exp_ns_kps_cmbd}{Expected numbers of kin-pairs, as output by
cmbn_exp_ns_kps. A matrix with rows for survey years and pairs, and columns
for kin-pair types}

\item{k}{Number of survey-years}

\item{p}{Sampling probability}
}
\value{
A matrix with rows for survey years and pairs, and columns for
kin-pair types
}
\description{
Find expected numbers of kin-pairs between sampled animals
}
