% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_obsd_gnpr_prbs.R
\name{find_exp_plods}
\alias{find_exp_plods}
\title{Find expected values of half-sibling versus unrelated pair PLODs given
various kinships}
\usage{
find_exp_plods(pgps, L)
}
\arguments{
\item{pgps}{Possible genopair probabilities given various kinships,
n_genotypes x n_genotypes x n_loci x n_kinships}

\item{L}{Number of genotype loci}
}
\value{
Named vector
}
\description{
Find expected values of half-sibling versus unrelated pair PLODs given
various kinships
}
