% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_obsd_gnpr_prbs.R
\name{find_hsp_up_plods}
\alias{find_hsp_up_plods}
\title{Find half-sibling versus unrelated pair PLODs}
\usage{
find_hsp_up_plods(oglps, L)
}
\arguments{
\item{oglps}{Observed genopair log-probabilities}

\item{L}{Number of genotype loci}
}
\value{
Vector with PLOD for each pair
}
\description{
Find half-sibling versus unrelated pair PLODs
}
