% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_obsd_gnpr_prbs.R
\name{find_oglps}
\alias{find_oglps}
\title{Find log-probabilities of observed genopairs given that the individuals have
a particular kinship.  Computes over batches of loci to limit memory usage to
~1Gb.}
\usage{
find_oglps(pgps, ind_gts, siips, L, sk = F)
}
\arguments{
\item{pgps}{Possible genopair probabilities given various kinships,
n_genotypes x n_genotypes x n_loci x n_kinships}

\item{ind_gts}{Individual genotypes. Genotypes of all individuals sampled}

\item{siips}{Sample-individual index pairs. The index of the individual that
each sample comes from, for each sample in all pairs of samples.}

\item{L}{Number of genotype loci}

\item{sk}{Boolean indicating that only a single kinship is considered}
}
\description{
Find log-probabilities of observed genopairs given that the individuals have
a particular kinship.  Computes over batches of loci to limit memory usage to
~1Gb.
}
