% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gpr_prbs.R
\name{find_pss_frst_gt_prbs}
\alias{find_pss_frst_gt_prbs}
\title{Find possible first genotype probabilities for genopairs, 3 x 3 x n_loci}
\usage{
find_pss_frst_gt_prbs(pss_gt_prbs, L)
}
\arguments{
\item{pss_gt_prbs}{Possible genotype probabilities}

\item{L}{Number of loci in each genotype}
}
\description{
Find possible first genotype probabilities for genopairs, 3 x 3 x n_loci
}
