% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gpr_prbs.R
\name{find_pss_gp_prbs_kps}
\alias{find_pss_gp_prbs_kps}
\title{Find possible genopair probabilities over multiple loci given multiple
kinships}
\usage{
find_pss_gp_prbs_kps(gts, L, knshps = c("HSP", "POP", "SP"))
}
\arguments{
\item{L}{Number of loci in each genotype}

\item{knshps}{Vector of codes representing kinships for which to calculate
possible genopair probabilities}
}
\value{
Array with dimensions (number of possible genotypes x number of
possible genotypes x number of loci x number of kinships)
}
\description{
Find possible genopair probabilities over multiple loci given multiple
kinships
}
