% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gpr_prbs.R
\name{find_pss_gp_prbs_pops}
\alias{find_pss_gp_prbs_pops}
\title{Find possible genopair probabilities for parent-offspring pairs}
\usage{
find_pss_gp_prbs_pops(ale_frqs, L, pss_frst_gt_prbs)
}
\arguments{
\item{ale_frqs}{2 x L allele frequencies matrix for each allele of each
possible genotype}

\item{L}{Number of loci in each genotype}

\item{pss_frst_gt_prbs}{Possible genotype probabilities for the first
genotype in each genopair}
}
\value{
Array with dimensions (number of possible genotypes x number of
possible genotypes x number of loci)
}
\description{
Find possible genopair probabilities for parent-offspring pairs
}
