% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pss_gnpr_prbs.R
\name{find_pss_gp_prbs_sps}
\alias{find_pss_gp_prbs_sps}
\title{Find possible genopair probabilities for self-pairs}
\usage{
find_pss_gp_prbs_sps(pss_frst_gt_prbs, L)
}
\arguments{
\item{pss_frst_gt_prbs}{Possible genotype probabilities for the first
genotype in each genopair}

\item{L}{Number of loci in each genotype}
}
\value{
Array with dimensions (number of possible genotypes x number of
possible genotypes x number of loci)
}
\description{
Find possible genopair probabilities for self-pairs
}
