% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gpr_prbs.R
\name{find_pss_gp_prbs_ups}
\alias{find_pss_gp_prbs_ups}
\title{Find possible genopair probabilities for unrelated pairs}
\usage{
find_pss_gp_prbs_ups(pss_gt_prbs)
}
\arguments{
\item{pss_gt_prbs}{Possible genotype probabilities}
}
\value{
Array (Number of possible genotypes x number of possible genotypes x
number of loci)
}
\description{
Find possible genopair probabilities for unrelated pairs
}
