% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pss_gnpr_prbs.R
\name{find_pss_gt_prbs}
\alias{find_pss_gt_prbs}
\title{Find possible genotype probabilities}
\usage{
find_pss_gt_prbs(ale_frqs)
}
\arguments{
\item{ale_frqs}{2 x L allele frequencies matrix for each allele of each
possible genotype}
}
\value{
Number of possible genotypes x number of loci matrix
}
\description{
Find possible genotype probabilities
}
