% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_features.R
\name{find_sim_yrs}
\alias{find_sim_yrs}
\title{Find simulation years}
\usage{
find_sim_yrs(fnl_yr, sim_len)
}
\arguments{
\item{fnl_yr}{Final year of simulation}

\item{sim_len}{Length of simulation}
}
\value{
Vector of years over which population will be simulated.
}
\description{
Find simulation years
}
