% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_obsd_gnpr_prbs.R
\name{find_smpl_gts}
\alias{find_smpl_gts}
\title{Find sample genotypes, extracted from matrix of individual genotypes,
n_samples x n_loci, rows ordered by survey-year then individual ID}
\usage{
find_smpl_gts(smpl_hsts, indl_gts)
}
\arguments{
\item{smpl_hsts}{Sample histories. Binary matrix with rows for animals and
columns for surveys.}

\item{indl_gts}{Individual genotypes (SNPs). Binary array with rows for
maternal and paternal alleles, columns for genotype loci, and another
dimension for animals.}
}
\value{
Binary array as for indl_gts but with the third dimension covering
all samples, including repeated samples of the same animals.
}
\description{
Find sample genotypes, extracted from matrix of individual genotypes,
n_samples x n_loci, rows ordered by survey-year then individual ID
}
