% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fnormalise}
\alias{fnormalise}
\title{Normalize a Numeric Vector to the 0-1 Range}
\usage{
fnormalise(x)
}
\arguments{
\item{x}{A numeric vector to be normalized.}
}
\value{
A numeric vector with values scaled between 0 and 1.
}
\description{
This function normalizes a numeric vector so that its values are scaled to lie within the 0 to 1 range.
If all elements in the vector are identical, the function returns a vector of ones.
}
