% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{identical_elements}
\alias{identical_elements}
\title{Check if All Elements are Identical}
\usage{
identical_elements(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Numeric vector to be checked for identical elements.}

\item{tol}{Tolerance level for equality comparison. Default is \code{.Machine$double.eps ^ 0.5}.}
}
\value{
\code{TRUE} if all elements are approximately equal; otherwise, \code{FALSE}.
}
\description{
This function checks if all elements in a numeric vector are approximately equal within a specified tolerance.
}
\examples{
values <- c(0.1, 0.2, 0.3, 0.1 + 0.2 + 0.3)
identical_elements(values)

}
