% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lin_interpolation}
\alias{lin_interpolation}
\title{Perform Linear Interpolation}
\usage{
lin_interpolation(xp, x0, x1, y0, y1)
}
\arguments{
\item{xp}{A numeric vector of new x values at which to interpolate.}

\item{x0}{A numeric vector of original x values (starting points for interpolation).}

\item{x1}{A numeric vector of original x values (ending points for interpolation).}

\item{y0}{A numeric vector of original y values corresponding to x0.}

\item{y1}{A numeric vector of original y values corresponding to x1.}
}
\value{
A numeric vector of interpolated y values corresponding to \code{xp}.
}
\description{
This function performs linear interpolation for a set of points.
}
\examples{
# Interpolate y at xp given the line segments (x0, y0) -> (x1, y1)
lin_interpolation(xp = c(1.5, 2.5),
                  x0 = c(1, 2), x1 = c(2, 3),
                  y0 = c(10, 20), y1 = c(20, 30))  # 15 25

}
