% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_exp_ns_kps.R
\name{make_exp_ns_kps_smpd_df}
\alias{make_exp_ns_kps_smpd_df}
\title{Make data frame for expected numbers of kin-pairs between sampled animals}
\usage{
make_exp_ns_kps_smpd_df(
  exp_N_t,
  s_yr_inds,
  phi,
  rho,
  lambda,
  alpha,
  srvy_yrs,
  k,
  p
)
}
\arguments{
\item{phi}{Individual survival rate}

\item{rho}{Per capita birth rate}

\item{lambda}{Population growth rate}

\item{alpha}{Age of sexual maturity}

\item{k}{Number of survey-years}

\item{p}{Sampling probability}
}
\value{
A data frame with rows for survey years and pairs, and columns for
kin-pair types
}
\description{
Make data frame for expected numbers of kin-pairs between sampled animals
}
