% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_pop.R
\name{plot_exp_pop}
\alias{plot_exp_pop}
\title{Plot expected population size over time}
\usage{
plot_exp_pop(sim_yrs, exp_N_t, ref_yr, exp_N_ref, srvy_yrs)
}
\arguments{
\item{sim_yrs}{Years over which population will be simulated}

\item{exp_N_t}{Expected population size in each year}

\item{ref_yr}{reference year at which expected population size is defined}

\item{exp_N_ref}{Expected population size in reference year}

\item{srvy_yrs}{Survey years}
}
\description{
Plot expected population size over time
}
\examples{
plot_exp_pop(
  sim_yrs = 1:20, exp_N_t = 20*1.05^(1:20), ref_yr = 1, exp_N_ref = 20,
  srvy_yrs = 20
)

}
