% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_obsd_gnpr_prbs.R
\name{plot_plods}
\alias{plot_plods}
\title{Plot half-sibling versus unrelated pair PLODs on log-scale}
\usage{
plot_plods(hsp_up_plods, exp_plods, show_legend = T)
}
\arguments{
\item{hsp_up_plods}{Half-sibling versus unrelated pair PLODs}

\item{exp_plods}{Expected values of half-sibling versus unrelated pair PLODs
for various kinships}
}
\value{
None
}
\description{
Plot half-sibling versus unrelated pair PLODs on log-scale
}
