% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prop_if}
\alias{prop_if}
\title{\code{prop_if} calls an other function}
\usage{
prop_if(x, na_rm = FALSE)
}
\arguments{
\item{x}{a data set}

\item{na_rm}{By default = TRUE}
}
\description{
\code{prop_if} calls an other function
}
