% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop_study.R
\name{sim_pop_study}
\alias{sim_pop_study}
\title{Simulate a population and a study of it}
\usage{
sim_pop_study(
  phi,
  lambda,
  N.init,
  hist.len,
  srvy.yrs,
  k,
  f.year,
  p,
  L,
  imaf,
  clvng.p,
  tmp.emgn,
  alpha,
  clvng.ints
)
}
\arguments{
\item{phi}{The individual survival rate}

\item{lambda}{The population growth rate}

\item{N.init}{Initial population size}

\item{hist.len}{Length of simulation}

\item{srvy.yrs}{Survey years}

\item{k}{Number of survey years}

\item{f.year}{Final year of simulation}

\item{p}{Capture probability}

\item{L}{Number of genotype loci}

\item{imaf}{Initial minor allele frequency}

\item{clvng.p}{Additional calving probability}

\item{tmp.emgn}{Rate of temporary immigration}

\item{alpha}{Age of sexual maturity}

\item{clvng.ints}{Whether females should be selected for calving by time
since last calving, producing intervals of standard length}
}
\value{
A dataframe with ID, mum, dad, capture, and calving columns, and
several attributes.
}
\description{
Simulate a population and a study of it
}
\examples{
# Need a global variable setting births to be stochastic
stch.bths = TRUE

# Simulate one population and study
pop_study = sim_pop_study(
  phi = 0.9, lambda = 1.05, N.init = 20, hist.len = 20, srvy.yrs = 20, k = 1,
  f.year = 20, p = 0.5, L = 10, imaf = 0.5, clvng.p = 0, tmp.emgn = 0,
  alpha = 5, clvng.ints = FALSE
)

# Look at it
head(pop_study)
names(attributes(pop_study))
}
