% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{to_agegrp}
\alias{to_agegrp}
\title{Generate age-group from age}
\usage{
to_agegrp(dt, grp_width = 5L, age_colname = "age",
  agegrp_colname = "agegrp", to_factor = TRUE, ...)
}
\arguments{
\item{dt}{A data.table with a column named \code{age}.}

\item{age_colname}{A string denoting the age column in \code{dt}.}

\item{to_factor}{A logical. If \code{TRUE}, then the age-groups
column is converted to factor.}

\item{...}{Pass arguements to \code{\link{agegrp_name}}.}

\item{colname}{A string denoting the name of the column that will be
created for age-groups.}
}
\value{
a data.table, invisibly.
}
\description{
\code{to_agegrp} creates a new column
}
\examples{
library(data.table)
library(CKutils)
to_agegrp(data.table(age = 0:99))[]
to_agegrp(data.table(age = 0:99), max_age = 80L)[]
to_agegrp(data.table(age = 0:99), grp_width = 10, max_age = 85)[]
}
