% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{validate_gamlss}
\alias{validate_gamlss}
\title{Stochastic prediction from a gamlss object}
\usage{
validate_gamlss(dtb, gamlss_obj, mc = 10L, orig_data = dtb)
}
\arguments{
\item{dtb}{A data.table from which come the observed variables}

\item{gamlss_obj}{A gamlss object}

\item{mc}{Number of Monte Carlo simulations, by default 10L}

\item{orig_data}{Initial data.table, defaults to \code{dtb}}
}
\value{
A data.table with observed and predicted variables
}
\description{
`validate_gamlss` returns a data.table with the observed and predicted
variable. Multiple predictions are drawn from the predicted
distributions. Useful for plotting with ggplot.
}
\examples{
\dontrun{
library(data.table)
library(gamlss)

dat <- data.table(agegrp = runif(100, 20, 60))
dat[, bmi := 25 + 0.05 * agegrp + rnorm(.N, 0, 2)]
mod <- gamlss(bmi ~ agegrp, sigma.fo = ~1,
  family = gamlss.dist::NO(), data = dat)

# mc Monte Carlo draws stacked under the observed rows ("type" column)
vg <- validate_gamlss(copy(dat), mod, mc = 5L, orig_data = copy(dat))
vg[, .N, by = type]
}
}
