% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{absorb_dt}
\alias{absorb_dt}
\title{Copy all columns of dt_i in dt_x}
\usage{
absorb_dt(dt_x, dt_i, on = ".NATURAL", exclude_col = NULL, verbose = FALSE)
}
\description{
Copy all columns of dt_i in dt_x
}
