% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKutils.R
\docType{package}
\name{CKutils}
\alias{CKutils-package}
\alias{CKutils}
\title{CKutils}
\description{
CKutils: A package with utility functions I use regularly.
}
\details{
This package contains some functions I use regularly when I develop simulation models.
}
\author{
Chris Kypridemos
}
\keyword{internal}
