% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{clamp}
\alias{clamp}
\title{dispatch fclamp or fclamp_int depending on input}
\usage{
clamp(x, a = 0, b = 1, inplace = FALSE)
}
\arguments{
\item{x}{Numeric vector to be clamped.}

\item{a}{Lower bound of the range. Default is 0.}

\item{b}{Upper bound of the range. Default is 1.}

\item{inplace}{If TRUE, perform the clamping in-place. Default is FALSE.}
}
\value{
A numeric vector with values clamped within the specified range.
}
\description{
This function clamps numeric values within a specified range, returning the result.
}
\examples{
library(CKutils)
values <- c(0.2, 0.8, 1.5, -0.5)
clamp(values, a = 0, b = 1)

}
