% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{clamp}
\alias{clamp}
\title{Clamp values to a specified range}
\usage{
clamp(x, a = 0, b = 1, inplace = FALSE)
}
\arguments{
\item{x}{A numeric vector of values to be clamped.}

\item{a}{A numeric value specifying the lower bound (default: 0).}

\item{b}{A numeric value specifying the upper bound (default: 1).}

\item{inplace}{Logical. If TRUE, the clamping operation is performed in-place; otherwise, a new vector is returned.}
}
\value{
A numeric vector with values clamped to the interval [a, b].
}
\description{
`clamp` limits the values in a numeric vector `x` so that all elements are constrained within the interval [a, b].
It dispatches to either `fclamp` for double precision values or `fclamp_int` for integer values.
}
