% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_if}
\alias{count_if}
\title{\code{count_if} calls an other function}
\usage{
count_if(x, na_rm = FALSE)
}
\arguments{
\item{x}{a data set}

\item{na_rm}{By default = TRUE}
}
\description{
\code{count_if} calls an other function
}
