% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{crossval_gamlss}
\alias{crossval_gamlss}
\title{Deterministic prediction from a gamlss object}
\usage{
crossval_gamlss(dtb, gamlss_obj, orig_data = dtb, colnam = "rank")
}
\arguments{
\item{dtb}{A data. table}

\item{gamlss_obj}{a gamlss object}

\item{orig_data}{original data.table}

\item{colnam}{column names}
}
\description{
`crossval_gamlss` returns the observed and predicted values of the dependent
 variable. Useful for cross-validation metrics.
}
