% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{csv_to_fst}
\alias{csv_to_fst}
\title{Convert CSV Files to FST Format}
\usage{
csv_to_fst(csv_files, compression = 100L, delete_csv = FALSE)
}
\arguments{
\item{csv_files}{A character vector of file paths to CSV files.}

\item{compression}{An integer specifying the compression level for the FST file. Default is 100L.}

\item{delete_csv}{Logical. If TRUE, deletes the original CSV file(s) after conversion. Default is FALSE.}
}
\value{
Invisibly returns NULL.
}
\description{
This function converts one or more CSV files to the FST file format. It reads each CSV file using \code{data.table::fread} and writes it as an FST file using \code{fst::write_fst}. Optionally, the original CSV file(s) can be deleted.
}
\examples{
\dontrun{
csv_to_fst(c("data.csv"), compression = 100L, delete_csv = TRUE)
}
}
