% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_ops.R
\name{detach_package}
\alias{detach_package}
\title{Detach a Loaded Package}
\usage{
detach_package(pkg)
}
\arguments{
\item{pkg}{Character string of package name. The name of the package to detach.}
}
\value{
Invisibly returns \code{TRUE} if the package was detached, or \code{FALSE} if it was not attached.
}
\description{
Detaches a loaded R package cleanly from the current session. This is useful
when reinstalling or updating a package during development.
}
\details{
If the specified package is attached multiple times (e.g., via multiple `library()` calls),
it will be fully detached. A message is displayed indicating whether the detachment occurred.
}
\examples{
\dontrun{
detach_package("ggplot2")
}

}
\seealso{
\code{\link{library}}, \code{\link{detach}}
}
