% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gpr_prbs.R
\name{find_pss_gp_prbs_hsps}
\alias{find_pss_gp_prbs_hsps}
\title{Find possible genopair probabilities for half-sibling pairs}
\usage{
find_pss_gp_prbs_hsps(pss_gp_prbs_ups, pss_gp_prbs_pops)
}
\arguments{
\item{pss_gp_prbs_pops}{}
}
\description{
Find possible genopair probabilities for half-sibling pairs
}
