% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pss_gnpr_prbs.R
\name{find_pss_gp_prbs_hsps}
\alias{find_pss_gp_prbs_hsps}
\title{Find possible genopair probabilities for half-sibling pairs}
\usage{
find_pss_gp_prbs_hsps(pss_gp_prbs_ups, pss_gp_prbs_pops)
}
\arguments{
\item{pss_gp_prbs_ups}{Possible genopair probabilities for unrelated pairs}

\item{pss_gp_prbs_pops}{Possible genopair probabilities for parent-offspring
pairs}
}
\value{
Array with dimensions (number of possible genotypes x number of
possible genotypes x number of loci)
}
\description{
Find possible genopair probabilities for half-sibling pairs
}
