% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pss_gnpr_prbs.R
\name{find_pss_gp_prbs_kps}
\alias{find_pss_gp_prbs_kps}
\title{Find possible genopair probabilities over multiple loci given multiple
close-kinships}
\usage{
find_pss_gp_prbs_kps(ind_gts, L, cks = c("HSP", "POP", "SP"))
}
\arguments{
\item{ind_gts}{Genotypes in 2 x L x n_individuals arrays, representing two
binary SNPs at each locus for each individual, excluding repeated samples
of the same individual in different surveys}

\item{L}{Number of loci in each genotype}

\item{cks}{Vector of codes representing kinships for which to calculate
possible genopair probabilities}
}
\value{
Array with dimensions (number of possible genotypes x number of
possible genotypes x number of loci x number of kinships)
}
\description{
Find possible genopair probabilities over multiple loci given multiple
close-kinships
}
