% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pss_gnpr_prbs.R
\name{find_pss_gp_prbs_ups}
\alias{find_pss_gp_prbs_ups}
\title{Find possible genopair probabilities for unrelated pairs}
\usage{
find_pss_gp_prbs_ups(pss_frst_gt_prbs)
}
\arguments{
\item{pss_frst_gt_prbs}{Possible genotype probabilities for the first
genotype in each genopair}
}
\value{
Array with dimensions (number of possible genotypes x number of
possible genotypes x number of loci)
}
\description{
Find possible genopair probabilities for unrelated pairs
}
