% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fquantile}
\alias{fquantile}
\title{Compute Quantiles Using Default R Type 7 Method}
\usage{
fquantile(x, probs, na_rm = TRUE)
}
\arguments{
\item{x}{A numeric vector from which quantiles are computed.}

\item{probs}{A numeric vector of probabilities for which quantiles are desired.}

\item{na_rm}{Logical flag indicating whether to remove NA values (default is true).}
}
\value{
A numeric vector containing the computed quantiles corresponding to the probabilities in \code{probs}.
}
\description{
This function computes quantiles for a numeric vector using the default R method (type 7).
It handles missing values and can remove them if requested.
}
