% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fquantile}
\alias{fquantile}
\title{\code{fquantile} calls an other function}
\usage{
fquantile(x, probs, na_rm = TRUE)
}
\arguments{
\item{x}{a data set}

\item{probs}{a probs object}

\item{na_rm}{By default = TRUE}
}
\description{
\code{fquantile} calls an other function
}
