% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fquantile_byid}
\alias{fquantile_byid}
\title{Compute Quantiles by Group}
\usage{
fquantile_byid(x, q, id, rounding = FALSE, na_rm = TRUE)
}
\arguments{
\item{x}{A numeric vector from which quantiles are computed.}

\item{q}{A numeric vector of probabilities for which quantiles are desired.}

\item{id}{A string vector representing group identifiers. The vector \code{x} must be sorted by \code{id}.}

\item{rounding}{Logical flag indicating whether to round the quantiles (default is false).}

\item{na_rm}{Logical flag indicating whether to remove NA values (default is true).}
}
\value{
A list where the first element is a vector of group IDs and subsequent elements are numeric vectors of computed quantiles for each group.
}
\description{
This function computes quantiles for subsets of a numeric vector defined by an ID vector.
The input vector must be sorted by the ID vector. Optionally, the results can be rounded.
}
