% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fqZABNB}
\alias{fqZABNB}
\title{Zero Adjusted Beta Negative Binomial Quantile Function}
\usage{
fqZABNB(p, mu, sigma, nu, tau, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{mu}{vector of positive means.}

\item{sigma}{vector of positive dispersion parameters.}

\item{nu}{vector of positive shape parameters.}

\item{tau}{vector of hurdle probabilities (0 < tau < 1).}

\item{lower_tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{log_p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
An integer vector of quantiles.
}
\description{
Quantile function for the Zero Adjusted (Hurdle) Beta Negative Binomial (ZABNB) distribution
with parameters mu (mean), sigma (dispersion), nu (shape), and tau (hurdle probability).
}
\details{
The zero adjusted (hurdle) beta negative binomial distribution has two parts:
a point mass at zero and a truncated BNB distribution for positive values.
}
\examples{
# Single values
fqZABNB(c(0.1, 0.5, 0.9), mu=2, sigma=1, nu=1, tau=0.1)

# Vector inputs with recycling
fqZABNB(c(0.25, 0.75), mu=c(1,2), sigma=0.5, nu=c(1,1.5), tau=0.1)

}
\references{
Rigby, R. A., Stasinopoulos, D. M., Heller, G. Z., and De Bastiani, F. (2019) 
Distributions for modelling location, scale, and shape: Using GAMLSS in R, 
Chapman and Hall/CRC.
}
