% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fqZISICHEL}
\alias{fqZISICHEL}
\title{Zero-Inflated Sichel Distribution Quantile Function}
\usage{
fqZISICHEL(p, mu, sigma, nu, tau, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{mu}{vector of positive means.}

\item{sigma}{vector of positive dispersion parameters.}

\item{nu}{vector of shape parameters (real values).}

\item{tau}{vector of zero-inflation parameters (0 < tau < 1).}

\item{lower_tail}{logical; if TRUE (default), probabilities are P[X <= x],
otherwise, P[X > x].}

\item{log_p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A numeric vector of quantiles.
}
\description{
Quantile function for the zero-inflated Sichel distribution with parameters 
mu (mean), sigma (dispersion), nu (shape), and tau (zero-inflation).
}
\details{
The zero-inflated Sichel distribution is a mixture of a point mass at zero 
and a (truncated at zero) Sichel distribution.
}
\examples{
# Vector inputs with recycling
fqZISICHEL(c(0.25, 0.75), mu=c(1,2), sigma=1, nu=-0.5, tau=0.1)

}
\references{
Rigby, R. A., Stasinopoulos, D. M., Heller, G. Z., and De Bastiani, F. (2019) 
Distributions for modeling location, scale, and shape: Using GAMLSS in R, 
Chapman and Hall/CRC.
}
