% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gnrt_folder_structure}
\alias{gnrt_folder_structure}
\title{Generate folder structure}
\usage{
gnrt_folder_structure(path = getwd())
}
\arguments{
\item{path}{A string scalar. The root folder where the folder structure will
be created.}
}
\description{
`gnrt_folder_structure` generates the folder structure for an IMPACTncd model
}
\details{
This is an auxilliary function: It creates the expected folder structure for
an IMPACTncd model.
}
