% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{guess_gamlss}
\alias{guess_gamlss}
\title{Prediction from a gamlss object in parallel}
\usage{
guess_gamlss(dtb, gamlss_obj, orig_data = gamlss_obj$data, nc = 1L)
}
\arguments{
\item{dtb}{A data.table}

\item{gamlss_obj}{gamlss object}

\item{orig_data}{original data.table}

\item{nc}{by default = 1L}
}
\description{
`guess_gamlss` returns a data.table with the predicted
 variable. `dtb` needs to have a column with percentiles named `rank_y`,
 where `y` the name of the predicted variable (i.e. bmi).
}
