% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{guess_polr}
\alias{guess_polr}
\title{Prediction from a MASS:polr object in parallel}
\usage{
guess_polr(dt, polr_obj)
}
\arguments{
\item{dt}{a data.table}

\item{polr_obj}{a polr object}
}
\value{
A data.table with the predicted variable.
}
\description{
`guess_polr` returns a data.table with the predicted
 variable. `dt` needs to have a column with percentiles named `rank_y`,
 where `y` the name of the predicted variable (i.e. active_days).
}
