% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{identical_elements}
\alias{identical_elements}
\title{Check if All Elements in a Numeric Vector Are Identical}
\usage{
identical_elements(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A numeric vector to test.}

\item{tol}{A numeric tolerance used for the comparison (default: .Machine$double.eps^0.5).}
}
\value{
A logical value indicating whether all elements in \code{x} are identical.
}
\description{
This function checks whether all elements in a numeric vector are equal within a specified tolerance.
It returns TRUE if all elements are identical (considering the tolerance), and FALSE otherwise.
}
