% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_dt.R
\name{is_valid_lookup_tbl}
\alias{is_valid_lookup_tbl}
\title{Check Validity of Lookup Table}
\usage{
is_valid_lookup_tbl(lookup_tbl, keycols, fixkey = FALSE)
}
\arguments{
\item{lookup_tbl}{The data.table representing the lookup table.}

\item{keycols}{A character vector specifying the key columns in the lookup table.}

\item{fixkey}{Logical. If TRUE, the function will automatically set the key of the lookup table to the provided key columns for best performance; default is FALSE.}
}
\value{
TRUE if the lookup table is valid; otherwise, an error is raised.
}
\description{
This function verifies that a lookup table meets the required structural conditions
for key-based lookups. It checks that key columns are of type integer (or factors stored as integers),
that each key column has a consecutive sequence of values, and that the table has the expected
number of rows based on all possible combinations of key values.
}
