% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_exp_ns_kps.R
\name{make_exp_ns_kps_smpd_df}
\alias{make_exp_ns_kps_smpd_df}
\title{Make data frame for expected numbers of kin-pairs between sampled animals}
\usage{
make_exp_ns_kps_smpd_df(srvy_yrs, exp_ns_kps_smpd)
}
\arguments{
\item{exp_ns_kps_smpd}{Expected numbers of kin-pairs between sampled animals,
as output by exp_ns_kps_smpd. A matrix with rows for survey years and
pairs, and columns for kin-pair types}
}
\value{
A data frame with row and column-names
}
\description{
Make data frame for expected numbers of kin-pairs between sampled animals
}
