% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{match_colnames_pattern}
\alias{match_colnames_pattern}
\title{Obtain matching names corresponding to patterns}
\usage{
match_colnames_pattern(dt, ...)
}
\arguments{
\item{dt}{A data.table from which the column names \code{names(dt)} or \code{colnames(dt)} will be tested}

\item{...}{A list of the names to match with the data.table ones. Needs to be made of character otherwise the function stops}
}
\description{
\code{match_colnames_pattern} returns the matching names of the argument \code{dt}
(i.e. \code{names(dt)}) corresponding to the regular expression patterns
provided. The patterns must be supported by \code{\link{grep}}.
This is based on \code{data.table:::patterns}
}
\examples{
library(data.table)
library(CKutils)
dt <- data.table(id = c("city", "year", "birth", "idp"), b = c("age", "year", "bp", "name"))
z <- list("id", "year", "b")
match_colnames_pattern(dt, z) #[1] "id" "b"
}
