% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{normalise}
\alias{normalise}
\title{normalise a vector to 0,1 range}
\usage{
normalise(x, ...)
}
\arguments{
\item{x}{A numeric vector to be normalized.}

\item{...}{Additional arguments to be passed to the underlying normalization function.}
}
\value{
A normalized version of the input numeric vector.
}
\description{
This function normalizes a numeric vector by checking for identical elements and applying a normalization function.
}
\examples{
library(CKutils)
x <- c(2, 4, 6, 8)
normalise(x)

}
