% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{outersect}
\alias{outersect}
\title{\code{outersect} Calculates the symmetrical set difference of subsets}
\usage{
outersect(x, y, ...)
}
\arguments{
\item{x, y}{vectors, data frames containing a sequence of items}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
A vector made of both contents from x and y, except from the duplicated items
}
\description{
\code{outersect} Calculates the symmetrical set difference of subsets
}
\examples{
x <- c("age", "year", "bp", "name")
y <- c("city", "year", "birth", "id")
outersect(x, y)
#"age" "bp" "name" "city" "birth" "id"
}
