% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_pop.R
\name{plot_exp_pop}
\alias{plot_exp_pop}
\title{Plot expected population size over time}
\usage{
plot_exp_pop(sim_years, exp_N_t, base_yr, exp_N_base, srvy_yrs)
}
\arguments{
\item{sim_years}{Years over which population will be simulated}

\item{exp_N_t}{Expected population size in each year}

\item{base_yr}{Base year at which expected population size is determined}

\item{exp_N_base}{Expected population size in base year}

\item{srvy_yrs}{Survey years}
}
\value{
NULL.
}
\description{
Plot expected population size over time
}
\examples{
plot_exp_pop(
  sim_years = 1:20, exp_N_t = 20*1.05^(1:20), base_yr = 1, exp_N_base = 20,
  srvy_yrs = 20
)

}
