% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prop_if}
\alias{prop_if}
\title{Proportion of TRUE Values in a Logical Vector}
\usage{
prop_if(x, na_rm = FALSE)
}
\arguments{
\item{x}{A logical vector.}

\item{na_rm}{Logical flag indicating whether to remove NA values before calculating the proportion (default is false).}
}
\value{
A numeric value representing the proportion of TRUE values in the vector.
}
\description{
This function calculates the proportion of TRUE values in a logical vector, with an option to remove missing values.
}
