% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_dt.R
\name{set_lookup_tbl_key}
\alias{set_lookup_tbl_key}
\title{Set Lookup Table Key}
\usage{
set_lookup_tbl_key(lookup_tbl, keycols)
}
\arguments{
\item{lookup_tbl}{A data.table whose key columns are to be set.}

\item{keycols}{A character vector specifying the key columns to be used.}
}
\value{
The original lookup table (invisibly) with its key set for efficient subsetting.
}
\description{
Sets the key columns of a lookup table to optimize performance for lookup operations.
The key columns are sorted (with a priority given to "year" if present) and then set
as the key for the data.table.
}
